apps/summarizer/backend/pyproject.toml
[tool.poetry]
name = "summarizer"
version = "0.1.0"
description = ""
authors = ["Diogo Schuarz <diogo.schuarz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asyncpg = "^0.29.0"
fastapi = "^0.115.0"
google-generativeai = "0.7.2"
langchain = "^0.3.0"
langchain-community = "^0.3.0"
langchain-core = "^0.3.5"
langchain-google-genai = "^2.0.0"
langchain-text-splitters = "^0.3.0"
langgraph = "^0.2.23"
langgraph-checkpoint = "^1.0.10"
psycopg-binary = "^3.2.2"
psycopg-pool = "^3.2.3"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
regex = "^2024.9.11"
requests = "^2.32.3"
sqlalchemy = "^2.0.35"
starlette = "^0.38.6"
streamlit = "^1.38.0"
tiktoken = "^0.7.0"
uvicorn = "^0.30.6"
wikipedia = "^1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
